simplify submodule loader + add test
